requirements.txt
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.11.7
botocore==1.14.7
certifi==2019.11.28
chardet==3.0.4
contractions==0.0.24
cssselect==1.1.0
docopt==0.6.2
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
gensim==3.8.1
idna==2.8
jieba3k==0.35.1
jmespath==0.9.4
joblib==0.14.1
lxml==4.4.2
newspaper3k==0.2.8
nltk==3.4.5
num2words==0.5.10
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
pyahocorasick==1.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-file==1.4.3
s3transfer==0.3.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
smart-open==1.9.0
soupsieve==1.9.5
textsearch==0.0.17
tinysegmenter==0.3
tldextract==2.2.2
Unidecode==1.1.1
urllib3==1.25.7
